move minIni out of tree
