+++ lib/minIni/src/minGlue-ffs.h
