+++ lib/minIni/src/wxMinIni.h
