+++ src/page/page_connections.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <pthread.h>
 #include <sys/time.h>
-#include "../minIni/minIni.h"
+#include "minIni.h"
 #include "../driver/esp32.h"
 #include "../core/common.hh"
 #include "../core/esp32_flash.h"
