+++ src/page/page_headtracker.c
@@ -1,6 +1,6 @@
 #include "page_headtracker.h"
 #include "page_common.h"
-#include "../minIni/minIni.h"
+#include "minIni.h"
 #include "style.h"
 #include "ht.h"
